[Chore] 테스트용 코드 분리
